Document the oops episode 2
